Extract Rust gatekeeper config module
